Changes required to use tt-mlir dialects out-of-tree (#2997)
### Problem description
While trying to set up an out-of-tree pass that lowers to TTKernel and
TTMetal dialects. Linking failed as it was not able to find the
definitions for MemorySpaceAttr, MetalLayoutAttr, TypeIDResolver, etc.


And another error while finding the TTMLIRConfig:
```
CMake Error at CMakeLists.txt:27 (find_package):
  Found package configuration file:

    ../tt-mlir/build/lib/cmake/ttmlir/TTMLIRConfig.cmake

  but it set TTMLIR_FOUND to FALSE so package "TTMLIR" is considered to be
  NOT FOUND.  Reason given by package:

  The following imported targets are referenced, but are missing:
  TTNN::TTNNOpModelLib
 ```

### What's changed
Export `TTNNOpModelLib` and `coverage_config` as part of `TTMLIR_EXPORTS`.
Set right dependencies for `MLIRTTKernelDialect` and `MLIRTTMetalDialect`