Update ttmetal.enqueue_program_op verifier
Update ttmetal.enqueue_program_op verifier to remove a check that
ensures that the number of buffer operands and the cb operands
are equal.

This need not be always true. e.g. a compute kernel might only
read and write to circular buffers and do not depend on any
global buffer.